requirements.txt
Flask=1.1.2
Flask-Cors==3.0.10
flask-smorest==0.30.0
marshmallow==3.11.1
flask-mongoengine==1.0.0
python-dotenv==0.17.0